src/test/java/com/parallelbrains/baselayout/controller/HomeControllerTest.java
package com.parallelbrains.baselayout.controller;

import com.parallelbrains.baselayout.controller.view.HomeController;
import junit.framework.Assert;
import org.junit.Test;
import org.springframework.ui.ExtendedModelMap;
import org.springframework.ui.Model;

public class HomeControllerTest {

	@Test
	public void testController() {
		HomeController controller = new HomeController();
		Model model = new ExtendedModelMap();
		Assert.assertEquals("home",controller.home(model));
		
		Object message = model.asMap().get("controllerMessage");
		Assert.assertEquals("This project is under development",message);
	}
}
